requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pyalex==0.18
python-dotenv==1.0.0
networkx==3.2.1
matplotlib==3.7.2
plotly==5.15.0
